# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Reservations API",
        "description": "API REST para gestão de hotéis, quartos e reservas (desafio técnico Foco Multimídia). A consulta de quartos é pública. Cadastro, edição e exclusão de quartos e todas as rotas de reservas exigem token: faça login em POST /login e use o botão Authorize.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local (Docker)"
        }
    ],
    "paths": {
...
        "/reserves": {
            "post": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Cria uma reserva",
                "description": "O total e as diárias são calculados pelo sistema a partir de daily_value. O pagamento é opcional e pode ser parcial, mas não pode ultrapassar o total. Formas de pagamento: 1 = dinheiro, 2 = pix, 3 = cartão.",
                "operationId": "b5c3af9f9d98784e08294b2ac7e196ec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "hotel_id",
                                    "room_id",
                                    "check_in",
                                    "check_out",
                                    "daily_value",
                                    "guests"
                                ],
                                "properties": {
                                    "hotel_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "room_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "check_in": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2027-01-10"
                                    },
                                    "check_out": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2027-01-13"
                                    },
                                    "daily_value": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "guests": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "name",
                                                "last_name",
                                                "phone"
                                            ],
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Maria"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Souza"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "5571999999999"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "payments": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "method",
                                                "value"
                                            ],
                                            "properties": {
                                                "method": {
                                                    "type": "integer",
                                                    "example": 2,
                                                    "enum": [
                                                        1,
                                                        2,
                                                        3
                                                    ]
                                                },
                                                "value": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 100
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva criada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Reserve"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "O quarto já está reservado em parte do período informado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Este quarto já está reservado de 10/01/2027 a 13/01/2027."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados inválidos: quarto de outro hotel, datas incorretas, check-in no passado, sem hóspede, forma de pagamento inválida ou pagamento maior que o total",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reserves/{id}": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Detalha uma reserva",
                "description": "Retorna hóspedes, diárias, pagamentos, saldo e status de pagamento.",
                "operationId": "0f769fd3ea8bb6f58890362ca9fa26d6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da reserva",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Reserve"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reserva não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Reserve": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 9
                    },
                    "hotel": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Hotel Foco Prime"
                            }
                        },
                        "type": "object"
                    },
                    "room": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Room 1 Hotel 1"
                            }
                        },
                        "type": "object"
                    },
                    "check_in": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-12-20"
                    },
                    "check_out": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-12-23"
                    },
                    "nights": {
                        "type": "integer",
                        "example": 3
                    },
                    "total": {
                        "type": "string",
                        "example": "300.00"
                    },
                    "paid": {
                        "type": "string",
                        "example": "100.00"
                    },
                    "balance": {
                        "type": "string",
                        "example": "200.00"
                    },
                    "payment_status": {
                        "type": "string",
                        "example": "parcial",
                        "enum": [
                            "pendente",
                            "parcial",
                            "quitado"
                        ]
                    },
                    "guests": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Maria"
                                },
                                "last_name": {
                                    "type": "string",
                                    "example": "Souza"
                                },
                                "phone": {
                                    "type": "string",
                                    "example": "5571999999999"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "dailies": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "date": {
                                    "type": "string",
                                    "format": "date",
                                    "example": "2026-12-20"
                                },
                                "value": {
                                    "type": "string",
                                    "example": "100.00"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "payments": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "method": {
                                    "type": "integer",
                                    "example": 2
                                },
                                "method_name": {
                                    "type": "string",
                                    "example": "pix"
                                },
                                "value": {
                                    "type": "string",
                                    "example": "100.00"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Cole apenas o token devolvido pelo login (sem a palavra Bearer).",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}